Initial import of (almost) all files from the coxeter_3.0_beta3.tgz archive
The import omitted 'modify_all' and 'trial' which seem to have no real usage.
Also all trailing whitespace was removed, and 'makefile' was renamed 'Makefile'

Finally the LICENSE now being in a separate file, its text was removed from
main.cpp, and references to it were modified to point to that file.

Currently the files do not compile; making this work will be the next commit.