Fix problem with dictionaries: sharing pointers led to multiple freeing
Dictionaries are tree structures for looking up stuff, which are used in the
'commands' unit to locate "commands", in the form of |CommandData| units, from
user input. Command completion of incomplete commands is implemented by linking
the incomplete (but unambiguous) input to the |CommandData| for the completed
command, by copying the pointer. But this means the dictionary has multiple
identical pointers, which it owns, which causes multiple freeing of those
pointers on destruction of the dictionary, which happens at program completion.
The solution made is to use |std::shared_ptr| for the payload pointers in the
dictionary.

It is curious that this programming error was not noticed when using Fokko's own
data types and allocation; apparently his memory arena scheme does not suffer
from multiple freeing, at least not when it happens at program termination. It
does cause a problem when |std::string| is used instead in the |CommandData|.