Fix compiler-dependent error that |sprintf| cannot accept |string::data|
The problem is that |std::string::data| returns a pointer to const char, even if
the string itself is no const. This changes in C++17, but rather than make such
a jump in C++ version, we just provide a pointer to the first character instead
of calling the |data| method; this works in all versions.

In passing we make the choice for C++11 explicit in the Makefile