Modify command mode initialization to use one less static variable
The command modes themselves were static pointers that were set to point to
static variables of initialization functions, which functions were only correct
because they were called only once, which was indeed so because those pointers
were static. If ever the initialization functions would be called more than
once, their own static |CommandTree| variable would be filled multiple times.

To make this a bit less brittle, the initialization functions are now passed as
argument to the constructor to be called from there; constructors are guaranteed
to be called only once for any object, whether static or not, so this is safe.
These initialization functions now no longer have any static variable.

As a supplementary benefit, the calls of |commandCompletion| can now be made
inside the constructor, after filling the tree, rather than having to be called
explicitly at the end of each command initialization function.