Simplify implementation of command completion
Knowing how command completion is implemented, the 'dictionary' implementation
need not use the |fullname| and |ambigPrefix| fields at all: all nodes with
absent action are ambiguous, and a node represents a complete command if it has
an action, and it either has no extension, or its action differs from that of
its first extension. This can be implemented at the 'dictionary' level, in a
method |has_own_action|, even though command completion is installed by the
client 'command' compilation unit.