Implement directory listing though forward iterator; import sl_list.h from Atlas
To be able neatly separate the dictionary implementation form the commands
module, the listing of directory entries in pre-order is implemented within the
directory module, rather than in the commands module using knowledge of the
internals of the directory structure. In order to do this a forward iterator is
provided, and in order to do that we incorporate singly linked list based stacks
from the Atlas project.