Remove all |using namespace| directives from header files, a mortal sin
This is a huge change, reflecting the insidiousness of the habit of writing many
|using namespace| directives (which inject one namespace into another with no
way of undoing the effect within the compilation unit seeing this) in the header
files. Since any client using a compilation unit is obliged to include its
header file, this practice leads to an nigh uncontrollable tsunami of cross
contamination of namespaces, utterly destroying the purpose of namespaces.

At on point Fokko warns to not do |using namespace kl| since apparently there
are definitions there that really don't like the cohabitation with other one.
During the editing for this commit we noticed some name coincidences that while
not apparently preventing compilation did lead to confusion while rectifying the
situation. Notably having a sub-namespace in the global namespace |commands|
that shares the name |interface| with a global namespace (its uncle so to speak)
makes it difficult to refer to that namespace within |commands|. Also kl.cpp
defines function |safeAdd| and |safeSubtract| in its anonymous namespace that
were in overloading competition with functions of the same name defined in
namespace |klsupport|; without the |using| directive for that namespace it had
to be decided case by case whether a local or a qualified name was appropriate.