Echo completed name when actively using command completion
Whenever one gives an incomplete command that the program automatically
completes (which includes the case of using autorepeat with an empty command)
the command that is actually being invoked is echoed back to the user for
clarity. This requires a bit of modification of the internal logic, notably
copying the shared pointer to the |CommandData| rather than just the action
pointer to implement autorepeat. As a consequence typing an empty command will
in certain circumstances where autorepeat is off result in showing a list of all
completions of the empty string; while it is a bit mysterious under which
conditions this happens exactly, it seems OK as a reaction.