Change implementation of |LaurentPolynomial<T>| to use |std::vector<T>|
This was one of the least used usages of |list::List<T>| so a good starting
point for phasing out the latter. It only really affects |MuPol|, uned in unequl
rank K-L computations only. The transition is not complete yet, in that some
further functions using this type need to be inspected. Also the use of vectors
can be much improved, it is still a lot allocate-then-fill based handling.

We should make use of |std:vector<long,containers::allocator>| instead.