Use |std::vector| for (all forms of) |HeckeMonomial|.
This runs pretty deep, since there are templated functions whose template type
is assumed to be one the the different (non templated) cases of |HeckeMonomial|,
which more or less forces the letter to all have similar implementations.

One sorting function |io::sortI| needed to bne given a |std::vector| overload, I
did not check if the older overload still has uses. By the way the
implementation of this sorting function is a curious form of three-way merge
sort using insertion sort to do the merge phase; Fokko does not tell where he
found this or whether he invented it himself. It's probably OK in effeciency.