Complete conversion of |KLContext::KLHelper| to pure "pointer to implementation"
All methods of |KLContext| are now immediately forwarded to |KLHelper|, avoiding
any back-and-forth between the two classes. Indeed, the pointer back from the
|KLHelper| class to its parent is now useless and was removed.

One advantage of this setup is that the assumptions that implementation methods
make can now (in future changes) be ensured by the methods of |KLContext|, by
automatically calling extension methods of external structures (like the
|KLSupport|) before calling those implementation methods, making the class less
brittle (dependent on conventions adhered to by client code).