Add |recursively_allocated| bitmap to track whether |standard_path| is set up
The problem was that with both |allocRowComputation| (which sets up the lists of
extremal elements all along the "standrd path|" to the identity) and a less
ambitious function |ensure_extr_row_exists| (which just adds a list at a given
element) being available, the |KLSupport| class cannot tell whether the presence
of a list at |y| implies the presence along the whole path. The added bitmap is
only set (and used) by the former function, so that it can safely cop out fast.