Remove abstract base class from |StandardSchubertContext|, and rename classes
So the abstract class |SchubertContext| becomes |AbstractSchubertContext| (and
unused) while |StandardSchubertContext| becomes|SchubertContext|, and is no
longer derived from any abstract class. This is probably the inverse change of
something Fokko did long ago with the intention of adding other implementations
of the abstract class, but in the absence of obvious ways to do that the old
setup is simply pointlessly heavy. Notably any change to the elaborate class
interface would also necessitate a corresponding change in the abstract class.

We left the abstract class in the code as historic documentation, for now.