Equip |compute_KL_row| with bitmap argument to avoid row completion testing
The recursive computation needs to ensure certain previous rows have been
completely computed before using their data; this can potentially lead to the
fairly expensive test that a row contains no undefined polynomials being called
many times for the same previous row. By maintaining a bitmap of rows for which
the current root invocation has already either completed the row or found the
row to be already complete, this repetition of work can be avoided, and this is
achieved without adding any extra permanent data to the |KLContext| class.