Improve efficiency of interval generation in |ClosureIterator::operator++|
The idea here is that iteration over a vector is much faster than over a bitmap
representing the same set, so while generating a new Bruhat interval, we
maintain a vector version of each interval, with the elements listed in order of
generation, so that traversal for new such intervals will be faster. In the
current version the only dynamic allocations during the traversal is a stack of
essentially |BitMap| values (for testing membership, not for iteration over);
since it has a known maximal depth this could implemented in an array for speed.

The same method of using both a bitmap and a vector of its elements was also
applied in |SchubertContext::closure|, which suffered from the same sluggishness
due to repeated iteration over a large but most of the time sparse bitset.