Make |SchubertContext::d_downset| field a |containers::vector<bitmap::BitMap>|
This allows the new method |down_set| to be fast, to the detriment of the old
|downset| method, which now needs to make a copy. But it will be phased out.
These changes also led to |schubert::ClosureIterator| to be rewritten.

However these changes make |ClosureIterator::operator++| very slow. The reason
is that we now iterate over a bitmap where we used to iterated ove the |List|
component of a |SubSet| which appeared to be redundant. Turns out Fokko have put
in this redundancy for a very good reason (but without documenting that reason
very explicitly, though |ClosureIterator::operator++| is said crucial for speed).