Convert |d_shift| from a list of row pointers to a single matrix.
For this purpose the Atlas |Matrix_base| class was imported into Coxeter as
|containers::matrix| in a somewhat simplified and modified form. Additional
methods are |push_row| and |grow|; the latter adds a number of rows filled with
identical entries. The row-orientation is more pronounced than in Atlas, notably
by exporting through |matrix::row| pointers to the beginnings of each row.