Remove the |SchubertContext::d_SubSet| field, replace by a method-local variable
This bitmap was only used by reference in |SchubertContext::extendContext|, and
its contents was reset on each such occasion, so keeping it around in the class
seems pointless. Nonetheless a bit of care was needed to make this change, as
resizing of the bitmap during extension happened rather indirectly in the
|SchubertContext::ContextExtension| constructor; that resizing must now be done
explicitly in |SchubertContext::fullExtension| instead.