Refine use of |Partition| class, introducing a |refine| method.
This method is not yet actually used, but the |cells::generalized_tau| function
template was reorganized to operate along the lines of refining all classes in
one pass while applying all applicable star operations in parallel. This meant
we now use a |multimap| with a vector value as key, and this led to the
detection of an error in containers.h, where the comparison type default was
that of image type, where it should be that of the key type; the error only was
noticeable when those two type are different (and led to a humongous message).