Review and revise |BitMap| implementation, making mixed capacities well defined
There are no longer restrictions on capacities in operations between |BitMap|
values: each operand represents the set of its bits 1, and the capacity of the
result is adapted if necessary. At the same time capacity handling is
independent of the actual contents of the operands: only their capacities are
used (so in particular the capacity is not decreased just to remove some bits
0). So for instance the operators & and &= result in a capacity the minimum of
the operand capacities, |, |=, ^, ^= the maximum, and the |andnot| method leaves
the capacity of the left operand unchanged. Shift operators increase or decrease
the capacity by the number that is the right hand operand.

We also distinguish |const_iterator| and |iterator| sub-classes, where the
latter has one extra method: |remove_target| which can clear the bit at which
the iterator points (without using the parent BitMap| explicitly); not used yet.