Fix an obvious oversight in |MinTable::insert|: variable too much localised.
The |MinNbr| local variable |r| needed to be initialised outside the loop, and
then assigned anew on every iteration. Instead there was no initialisation and
the assigment in the loop was made to the variable definition; the fact that the
defining expression already used the (then uninitialised) variable was signalled
as suspect by the (optimizing) compiler. This is fixed now.